Added error in case there is no user by the required name for store policies promotion

+++ SadnaSrc/SadnaSrc/UserSpot/UserPolicyService.cs
@@ -25,6 +25,10 @@ namespace SadnaSrc.UserSpot
         }       
         public static void PromoteStorePolicies(string userName,string store,StoreManagerPolicy.StoreAction[] actionsToAdd)
         {
+            if (_userDL.IsUserNameExist(userName))
+            {
+                throw new UserException(MarketError.DbError, "No user by the name '" + userName + " has been found for promotion!");
+            }
             foreach (StoreManagerPolicy.StoreAction oldAction in Enum.GetValues(typeof(StoreManagerPolicy.StoreAction)))
             {
                 _userDL.DeleteUserStorePolicy(userName, new StoreManagerPolicy(store, oldAction));
